pyproject.toml
[tool.poetry]
name = "keycloak_extend"
version = "0.0.1"
description = "Python keycloak extension for python-keycloak package."
license = "MIT"
readme = "README.md"
keywords = [ "keycloak", "openid", "oidc" ]
authors = [
    "BioticsAI",
]
packages = [
    { include = "keycloak_extend", from = "src/" },
    { include = "keycloak_extend/**/*.py", from = "src/" },
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = ">=2.20.0"
requests-toolbelt = ">=0.6.0"
deprecation = ">=2.1.0"
jwcrypto = ">=1.5.4"
httpx = ">=0.23.2"
async-property = ">=0.2.2"
python-keycloak = "^4.6.2"


[tool.poetry.dev-dependencies]
tox = ">=4.0.0"
pytest = ">=7.1.2"
pytest-cov = ">=3.0.0"
pytest-asyncio = ">=0.23.7"
wheel = ">=0.38.4"
pre-commit = ">=3.5.0"
isort = ">=5.10.1"
black = ">=22.3.0"
flake8 = ">=7.0.0"
flake8-docstrings = ">=1.6.0"
commitizen = ">=2.28.0"
cryptography = ">=42.0.0"
codespell = ">=2.1.0"
darglint = ">=1.8.1"
twine = ">=4.0.2"
freezegun = ">=1.2.2"
docutils = "<0.21"
pytest-docker = "^3.1.1"

[tool.poetry.extras]
docs = [
    "mock",
    "alabaster",
    "commonmark",
    "recommonmark",
    "sphinx",
    "sphinx-rtd-theme",
    "readthedocs-sphinx-ext",
    "m2r2",
    "sphinx-autoapi",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
markers = [
    "unit: mark as unit test, runs outside docker container",
    "integration: mark as integration test, runs in docker container",
]